// src/app/common/model/org-model.ts
export class AddCompany {
  companyName?: any;
  companyId?: any;
  parentId?: any;
  regNo?: any;
  legalPerson?: any;
  telNumber?: any;
  email?: any;
  fax?: any;
  zipCode?: any;
  foundDate?: any;
  address?: any;
  scale?: any;
  category?: any;
  introduction?: any;
}


export class AddDepart {
  deptName?: any;
  companyId?: any;
  parentId?: any;
}